検索結果画面のレイアウトを作成した
