--- .fvm/flutter_sdk
@@ -1,1 +1,0 @@
-/Users/uemurayuusuke/fvm/versions/stable
