+++ .gitignore
@@ -324,3 +324,6 @@ lint/tmp/
 *.hprof
 
 /.metadata
+
+
+.fvm/flutter_sdk
